Move to a submodule method
This commit moves the clap juce extensions to a model where both
they have clap as a submodule and they may be used as a submodule.
At the same time we move surge to use the submodule.

Changes are
- Submodule clap
- Update Documentation

Additionally, we resolve an issue with buscount and non eanbled
busses with a quick workaround for now while we figure out how
to properly deal with bus handshake