A Change for Destructive Value Set
When modulation arrived in bitwig as a destructive value
set, the param update would feedback and cause a runaway
modulation. This impacted Dexed but I bet it impacts
other plugins also.

This fix makes sure that when handling an event from
the clap queue, we do not re-send it out the clap queue,
but allows any other JUCE listener to continue to act.

It also allows a best-efforts string value for a param
for juce plugins using the legacy parameter methods.