Rename some extension decoration classes
Based on user feedback the name `clap_juce_extensions::clap_extensions`
was terribly confusing. Sounded like a way to extend juce to access
clap extensions. So we renamed it to `clap_juce_extensions::clap_juce_audio_processor_capabilities`
since it is extensions to juce for clap which augments your audio processor
with capabilities (and make a similar change to params).

This is a breaking change if you use extensions but is fixed with
a simple rename.