STEPPED parameters mis-handled (#148)
AudioParameterInt and CHoice use 0...1 normalizationfor their
lists. This makes them error prone under range changes, which is
an overall JUCE problem, but also means if we indicate them
as STEPPED in clap you can only select the extream in the bitwig
params view and so on. So for now stay with JUCE and make
these discrete types not stepped.