Fix a parameter threading error; Juce8 in CI; CPM upgrade
I had mistakenly thought htat the value change listeners on param
were not audio safe, so moved them to the main thread. This was
wrong, and this commit corrects that.

While in there add a Juce 8 test, which restructured CI a smidge
and also upgrade CPM since it throws a warning with newer cmakes